requirements.txt
fastapi
requests
google-cloud-logging
